pyproject.toml
[project]
name = "basetrader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.3",
    "asyncio>=4.0.0",
    "py-clob-client>=0.34.5",
    "python-dotenv>=1.2.1",
    "web3>=7.14.0",
    "websockets>=15.0.1",
]

[dependency-groups]
dev = [
    "ruff>=0.14.14",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.23.0",
]
